调整首页代码结构
