+++ tdesign-component/example/lib/base/example_route.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import '../about.dart';
-import '../main_run.dart';
+import '../config.dart';
+import '../home.dart';
 import 'api_widget.dart';
 import 'example_base.dart';
 import 'example_widget.dart';
