+++ tdesign-component/example/lib/web/web.dart
@@ -8,7 +8,8 @@ import 'package:tdesign_flutter/tdesign_flutter.dart';
 
 import '../base/api_widget.dart';
 import '../base/example_base.dart';
-import '../main_run.dart';
+import '../config.dart';
+import '../home.dart';
 import 'code_widget.dart';
 
 class WebMainBody extends StatefulWidget {
