+++ tdesign-component/example/test/widget_test.dart
@@ -10,8 +10,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:tdesign_flutter/tdesign_flutter.dart';
 import 'package:tdesign_flutter_example/base/web_md_tool.dart';
+import 'package:tdesign_flutter_example/config.dart';
 
-import 'package:tdesign_flutter_example/main_run.dart';
+import 'package:tdesign_flutter_example/home.dart';
+import 'package:tdesign_flutter_example/main.dart';
 
 void main() async {
